Adding Authentication to Journal Endpoints

+++ src/main/java/net/engineeringdigest/journalApp/Service/JournalEntryService.java
@@ -16,7 +16,6 @@ import org.springframework.transaction.annotation.Transactional;
 
 @Component
 public class JournalEntryService {
-    private static final Logger log = LoggerFactory.getLogger(JournalEntryService.class);
 
     @Autowired
     private JournalEntryRepository journalEntryRepository;
@@ -32,7 +31,7 @@ public class JournalEntryService {
             journalEntry.setDate(LocalDateTime.now());
             JournalEntry saved = journalEntryRepository.save(journalEntry);
             user.getJournalEntries().add(saved);
-            userService.saveEntry(user);
+            userService.saveUser(user);
         } catch (Exception e){
             System.out.println(e);
             throw new RuntimeException("An error occured while saving the entry.",e);
@@ -55,11 +54,21 @@ public class JournalEntryService {
         return journalEntryRepository.existsById(id);
     }
 
-    public void deleteById(ObjectId id, String userName){
-//        userService.removeJournalEntryFromAllUsers(id);
-        User user = userService.findByUserName(userName);
-        user.getJournalEntries().removeIf(x->x.getId().equals(id));
-        userService.saveEntry(user);
-        journalEntryRepository.deleteById(id);
+    @Transactional
+    public boolean deleteById(ObjectId id,String userName){
+        boolean removed = false;
+        try {
+            User user = userService.findByUserName(userName);
+            removed = user.getJournalEntries().removeIf(x->x.getId().equals(id));
+            if(removed){
+                userService.saveUser(user);
+                journalEntryRepository.deleteById(id);
+            }
+        } catch (Exception e) {
+            System.out.println(e);
+            throw new RuntimeException("An error occured while deleting the entry.", e);
+        }
+        return removed;
     }
+
 }
